Update_1.2
Cart logic completed;

+++ src/controllers/cartController.js
@@ -5,9 +5,24 @@ import {
   deleteProductFromCart,
 } from "../services/cartService.js"
 
+// Util for validation of cart input
+const validateCartInput = (userId, medicine_id, quantity) => {
+  if (!userId || !medicine_id || quantity === undefined) {
+    throw new Error("User, product, and quantity must be specified");
+  }
+
+  if (typeof medicine_id !== "number" || typeof quantity !== "number") {
+    throw new Error("Product identifier and quantity must be integers");
+  }
+};
+
+
 const listCart = async (req, res, next) => {
   try {
     const userId = req.user.id;
+
+    if (!userId) throw new Error("User Required");
+
     const card = await getCart(userId);
 
     res.status(200).json(card);
@@ -19,39 +34,45 @@ const listCart = async (req, res, next) => {
 const addToCart = async (req, res, next) => {
   try {
     const userId = req.user.id;
-    const { productId, quantity } = req.body;
+    const { medicine_id, quantity } = req.body;
 
-    const result = await addProductToCart(userId, productId, quantity);
+    validateCartInput(userId, medicine_id, quantity); 
 
-    res.status(201).json(result);
+    await addProductToCart(userId, medicine_id, quantity);
+
+    res.status(201).json({message: "Product added successfully"});
   } catch (error) {
     next(error);
   }
 };
 
-const updateCart = async (req, res) => {
+const updateCart = async (req, res, next) => {
   try {
     const userId = req.user.id;
-    const { productId, quantity } = req.body;
+    const { medicine_id, quantity } = req.body;
 
-    await updateCartProducts(userId, productId, quantity);
+    validateCartInput(userId, medicine_id, quantity); 
+
+    await updateCartProducts(userId, medicine_id, quantity);
 
     res.status(200).json({ message: "Card updated successfully" })
   } catch (error) {
-    res.status(500).json({ message: `${error.message}` });
+    next(error);
   }
 };
 
-const removeFromCart = async (req, res) => {
+const removeFromCart = async (req, res, next) => {
   try {
     const userId = req.user.id;
-    const { productId } = req.params.id;
+    const productId = Number(req.params.id);
 
-    await deleteProductFromCart(userId, productId);
+    if (!userId || !productId) throw new Error("User and product must be specified");
+
+    await deleteProductFromCart(userId, productId); 
 
     res.status(200).json({ message: "Product removed successfully" });
   } catch (error) {
-    res.status(500).json({ message: `${error.message}` });
+    next(error);
   }
 };
